test(client-streams): versioned-table structural + byte-exact changelog + behavioral golden parity (KIP-889/914)
- wire: add message.timestamp.type=CreateTime to versioned changelog config
- builder: REUSE_KTABLE_SOURCE_TOPICS for versioned tables (changelog=source topic)
- test_driver: retain+expose drained changelog (drain_changelog)
- goldens: switch capture to Serdes.Long (matches I64Serde), drop unsupported source tombstone

+++ crates/client-streams/tests/jvm-capture/src/main/java/crabka/capture/Capture.java
@@ -771,12 +771,12 @@ public final class Capture {
     static Topology versionedTable() {
         StreamsBuilder b = new StreamsBuilder();
         b.table("in",
-                Consumed.with(Serdes.String(), Serdes.Integer()),
-                Materialized.<String, Integer>as(
+                Consumed.with(Serdes.String(), Serdes.Long()),
+                Materialized.<String, Long>as(
                         Stores.persistentVersionedKeyValueStore("vt", Duration.ofMillis(600_000)))
-                    .withKeySerde(Serdes.String()).withValueSerde(Serdes.Integer()))
+                    .withKeySerde(Serdes.String()).withValueSerde(Serdes.Long()))
             .toStream()
-            .to("out", Produced.with(Serdes.String(), Serdes.Integer()));
+            .to("out", Produced.with(Serdes.String(), Serdes.Long()));
         return b.build(optimizedProps());
     }
 
@@ -787,12 +787,12 @@ public final class Capture {
     static Topology versionedTableUnoptimized() {
         StreamsBuilder b = new StreamsBuilder();
         b.table("in",
-                Consumed.with(Serdes.String(), Serdes.Integer()),
-                Materialized.<String, Integer>as(
+                Consumed.with(Serdes.String(), Serdes.Long()),
+                Materialized.<String, Long>as(
                         Stores.persistentVersionedKeyValueStore("vt", Duration.ofMillis(600_000)))
-                    .withKeySerde(Serdes.String()).withValueSerde(Serdes.Integer()))
+                    .withKeySerde(Serdes.String()).withValueSerde(Serdes.Long()))
             .toStream()
-            .to("out", Produced.with(Serdes.String(), Serdes.Integer()));
+            .to("out", Produced.with(Serdes.String(), Serdes.Long()));
         return b.build();
     }
